refactor: move toast to components
