+++ site/src/features/pages/home/components/page.tsx
@@ -2,10 +2,10 @@ import type { GetTranslatedValue } from '@junat/core/i18n'
 import type { LocalizedStation } from '~/lib/digitraffic'
 import type { Locale } from '~/types/common'
 
-import React from 'react'
 import dynamic from 'next/dynamic'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
+import React from 'react'
 
 import { SITE_NAME } from '@junat/core/constants'
 import { getAccuracyWithUnit } from '@junat/core/geolocation'
@@ -16,9 +16,9 @@ import HeartFilled from '~/components/icons/heart_filled.svg'
 import List from '~/components/icons/list.svg'
 import { Notification } from '~/components/notification'
 import { StationList } from '~/components/station_list'
+import { useToast } from '~/components/toast'
 import { ToggleButton } from '~/components/toggle_button'
 import { SearchBar } from '~/features/search'
-import { useToast } from '~/features/toast'
 import { useClientStore } from '~/hooks/use_client_store'
 import { useFavorites } from '~/hooks/use_favorites'
 import { useLocale, useTranslations } from '~/i18n'
