+++ site/src/features/pages/single_train/components/page.tsx
@@ -18,7 +18,7 @@ import Calendar from '~/components/icons/calendar.svg'
 import ObjectHorizontalLeft from '~/components/icons/object_horizontal_left.svg'
 import Share from '~/components/icons/share.svg'
 import { Spinner } from '~/components/spinner'
-import { useToast } from '~/features/toast'
+import { useToast } from '~/components/toast'
 import { useLocale, useTranslations } from '~/i18n'
 import Page from '~/layouts/page'
 import interpolateString from '~/utils/interpolate_string'
