--- site/src/features/toast/index.ts
@@ -1,6 +1,0 @@
-export { useToast } from './stores/toast'
-
-export { Provider as ToastProvider } from './components/provider'
-export { Toast } from './components/toast'
-
-export type { ToastProps } from './components/toast'
